+++ internal/datasource/constants.go
@@ -1,6 +1,5 @@
 package datasource
 
-const sliceStringProp string = "slice_string"
 const extraTokensProp string = "extra_tokens"
 const nameProp string = "name"
 const typeProp string = "type"
